requirements.txt
# Core trading libraries
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
ta==0.10.2

# Async and scheduling
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# AI/ML
tensorflow==2.15.0
scikit-learn==1.3.2
stable-baselines3==2.1.0

# Technical analysis
talib-binary==0.4.26

# Notifications
python-telegram-bot==20.7

# Configuration and logging
python-dotenv==1.0.0
loguru==0.7.2

# Web framework for monitoring
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
pydantic==2.5.0
asyncio==3.4.3
aiohttp==3.9.1
